:bug: 빌드 실패 현상 해결
결국 `pages`는 순수하게 페이지만 있어야 최적화가 가능하다고 한다. https://nextjs.org/docs/messages/page-without-valid-component
