+++ pages/about/index.tsx
@@ -1,11 +1,19 @@
 import styled from '@emotion/styled';
 import React, { useRef } from 'react';
 
-import Introduction from './Introduction';
 import { GradientType } from '@components/Metaball/types';
+import TextStyle from '@components/Text';
 
 import useMetaball from '@hooks/useMetaball';
 import useWindow from '@hooks/useWindow';
+import { ForwardedCanvas } from '@components/Metaball';
+import TransitionText from '@components/Text/TransitionText';
+
+enum IntroductionEnum {
+  Maticulous,
+  Documentation,
+  Curious,
+}
 
 /** TODO:
  * @todo
@@ -61,6 +69,237 @@ const AboutPage = () => {
       ></Introduction.Curious>
     </>
   );
+};
+
+const Styled = {
+  Introduction: styled.section`
+    position: relative;
+    height: 1000px;
+    opacity: 1;
+  `,
+  TextContainer: styled.div`
+    position: relative;
+    transform: translate3d(100px, 300px, 0);
+  `,
+};
+
+const getGradientColors = (key: string): [GradientType, GradientType] => {
+  const bgGradients = ['#ab0746', '#943bc3', '#390fa3', '#16083e'];
+  const metaballGradients = ['#ff3dbb', '#ffb151', '#ffbeb5', '#ffbf00'];
+
+  const index = Object.keys(Introduction).indexOf(key);
+
+  return [
+    [bgGradients[index], bgGradients[index + 1]],
+    [metaballGradients[index], metaballGradients[index + 1]],
+  ];
+};
+
+const Maticulous = ({ width, height }: { width: number; height: number }) => {
+  const [initialGradientColors, metaballGradientColors] = getGradientColors(
+    IntroductionEnum[0]
+  );
+
+  const canvasRef = useRef<HTMLCanvasElement | null>(null);
+
+  useMetaball({
+    canvasRef: canvasRef,
+    gradient: initialGradientColors,
+    metaballGradient: metaballGradientColors,
+    mainMetaball: {
+      x: (width / 100) * 90,
+      y: (height / 100) * 25,
+      r: 200,
+    },
+    staticBubbles: [
+      {
+        x: (width / 100) * 90,
+        y: (height / 100) * 25,
+        r: 300,
+        v: [-14, 14.2],
+        to: {
+          x: (width / 100) * 90 - 400,
+          y: (height / 100) * 25 + 400,
+        },
+      },
+    ],
+    options: {
+      bubbleNum: 0,
+      absorbBallNum: 0,
+      canvasWidth: width,
+      canvasHeight: height,
+    },
+  });
+
+  return (
+    <Styled.Introduction>
+      <ForwardedCanvas
+        width={width}
+        height={height}
+        ref={canvasRef}
+      ></ForwardedCanvas>
+
+      <Styled.TextContainer>
+        <TransitionText $delay={1} $size="xxl" $fontWeight="extrabold">
+          <TextStyle.MainCopy>
+            <div>코드에는</div>
+            <strong>이야기</strong>가 있다고 믿어요
+          </TextStyle.MainCopy>
+
+          <TextStyle.Default>
+            <div>개발을 하는 모든 과정 자체를</div>
+            <div>유저와 동료에게 이야기한다고 생각해요.</div>
+            <div>사소한 UX 오류라도 개선하려 노력하며</div>
+            <div>주석 하나라도 애정을 담아 작성해요.</div>
+            <div>유저에게도, 동료에게도 좋은 이야기만 들려주고 싶은</div>
+            <div>개발자이자 이야기꾼입니다.</div>
+          </TextStyle.Default>
+        </TransitionText>
+      </Styled.TextContainer>
+    </Styled.Introduction>
+  );
+};
+
+const Documentation = ({
+  width,
+  height,
+}: {
+  width: number;
+  height: number;
+}) => {
+  const [initialGradientColors, metaballGradientColors] = getGradientColors(
+    IntroductionEnum[1]
+  );
+
+  const canvasRef = useRef<HTMLCanvasElement | null>(null);
+
+  useMetaball({
+    canvasRef: canvasRef,
+    gradient: initialGradientColors,
+    metaballGradient: metaballGradientColors,
+    mainMetaball: {
+      x: (width / 100) * 90,
+      y: (height / 100) * 25,
+      r: 200,
+    },
+    staticBubbles: [
+      {
+        x: (width / 100) * 90,
+        y: (height / 100) * 25,
+        r: 300,
+        v: [-14, 14.2],
+        to: {
+          x: (width / 100) * 90 - 400,
+          y: (height / 100) * 25 + 400,
+        },
+      },
+    ],
+    options: {
+      bubbleNum: 0,
+      absorbBallNum: 0,
+      canvasWidth: width,
+      canvasHeight: height,
+    },
+  });
+
+  return (
+    <Styled.Introduction>
+      <ForwardedCanvas
+        width={width}
+        height={height}
+        ref={canvasRef}
+      ></ForwardedCanvas>
+
+      <Styled.TextContainer>
+        <TransitionText $delay={1} $size="xxl" $fontWeight="extrabold">
+          <TextStyle.MainCopy>
+            <strong>문서화</strong>를 지향해요
+          </TextStyle.MainCopy>
+
+          <TextStyle.Default>
+            <div>아무리 좋은 이야기라도</div>
+            <div>줄거리 요약보다 이해하기 빠를 수 없다 생각해요.</div>
+            <div>제가 구현했던, 생각했던 모든 과정들을</div>
+            <div>기록으로 남겨놓아요.</div>
+          </TextStyle.Default>
+        </TransitionText>
+      </Styled.TextContainer>
+    </Styled.Introduction>
+  );
+};
+
+const Curious = ({ width, height }: { width: number; height: number }) => {
+  const [initialGradientColors, metaballGradientColors] = getGradientColors(
+    IntroductionEnum[2]
+  );
+
+  const canvasRef = useRef<HTMLCanvasElement | null>(null);
+
+  useMetaball({
+    canvasRef: canvasRef,
+    gradient: initialGradientColors,
+    metaballGradient: metaballGradientColors,
+    mainMetaball: {
+      x: (width / 100) * 90,
+      y: (height / 100) * 25,
+      r: 200,
+    },
+    staticBubbles: [
+      {
+        x: (width / 100) * 90,
+        y: (height / 100) * 25,
+        r: 300,
+        v: [-14, 14.2],
+        to: {
+          x: (width / 100) * 90 - 400,
+          y: (height / 100) * 25 + 400,
+        },
+      },
+    ],
+    options: {
+      bubbleNum: 0,
+      absorbBallNum: 0,
+      canvasWidth: width,
+      canvasHeight: height,
+    },
+  });
+
+  return (
+    <Styled.Introduction>
+      <ForwardedCanvas
+        width={width}
+        height={height}
+        ref={canvasRef}
+      ></ForwardedCanvas>
+
+      <Styled.TextContainer>
+        <TransitionText $delay={1} $size="xxl" $fontWeight="extrabold">
+          <TextStyle.MainCopy>
+            <strong>호기심</strong>이 많아요
+          </TextStyle.MainCopy>
+
+          <TextStyle.Default>
+            <div>재밌는 것을 보면 따라해요!</div>
+            <div>물론 막히면 스트레스를 엄~청 많이 받지만</div>
+            <div>구현했을 때 짜릿함 하나만으로도</div>
+            <div>개발을 좋아하는 개발자입니다.</div>
+          </TextStyle.Default>
+        </TransitionText>
+      </Styled.TextContainer>
+    </Styled.Introduction>
+  );
+};
+
+/**
+ * @description
+ * Introduction이라는 말도 전체를 다 표현하지 못한다는 생각에 이런 식으로 작성했다.
+ * 이것이 Introduction을 위한 하위 컴포넌트임을 분명히 하기 위해 몇 시간을 고민한 끝에, 약간 DOM보다 좀 더 upgrade된 방법을 사용했다.
+ * 어쩌면 누군가는 욕할 수 있겠지만, 이게 좀 더 깔끔하게 관리하는 듯하다 나는.
+ */
+const Introduction = {
+  Maticulous,
+  Documentation,
+  Curious,
 };
 
 export default AboutPage;
